backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
pydantic
requests
gitpython
markdown
tiktoken
PyYAML
openai